refactor(components): cleans structure
