--- src/components/SuccessAnimation.module.css
@@ -1,39 +1,0 @@
-.container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 9999;
-  animation: fadeIn 0.3s ease-in;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-.animationWrapper {
-  width: 100%;
-  height: 100%;
-  max-width: 600px;
-  max-height: 600px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  font-family: "Manrope", sans-serif;
-}
-
-.animation {
-  width: 100%;
-  height: 100%;
-}
