--- src/components/create/steps/index.ts
@@ -1,4 +1,0 @@
-export { StepBasics } from "./StepBasics";
-export { StepParties } from "./StepParties";
-export { StepEvidence } from "./StepEvidence";
-export { StepReview } from "./StepReview";
