--- src/components/profile/index.ts
@@ -1,4 +1,0 @@
-// Barrel exports for profile components
-export { AddContactDialog } from "./AddContactDialog";
-export { PendingPaymentsDialog } from "./PendingPaymentsDialog";
-export { PendingExecutionsDialog } from "./PendingExecutionsDialog";
