--- src/components/profile/views/index.ts
@@ -1,3 +1,0 @@
-export { ProfileOverview } from "./ProfileOverview";
-export { ContactsView } from "./ContactsView";
-export { SettingsView } from "./SettingsView";
